调整mdsclient rpc重试逻辑

+++ test/tools/etcd_client_test.cpp
@@ -15,7 +15,6 @@ class EtcdClientTest : public ::testing::Test {
  protected:
     void SetUp() {
         system("rm -fr test1.etcd");
-        system("rm -fr test2.etcd");
         etcdPid = ::fork();
         if (0 > etcdPid) {
             ASSERT_TRUE(false);
@@ -30,13 +29,17 @@ class EtcdClientTest : public ::testing::Test {
             ASSERT_EQ(0, execl("/bin/sh", "sh", "-c", runEtcd.c_str(), NULL));
             exit(0);
         }
+        running = true;
     }
 
     void TearDown() {
-        system(("kill " + std::to_string(etcdPid)).c_str());
-        std::this_thread::sleep_for(std::chrono::seconds(2));
+        if (running) {
+            system(("kill " + std::to_string(etcdPid)).c_str());
+            std::this_thread::sleep_for(std::chrono::seconds(2));
+        }
     }
 
+    bool running;
     pid_t etcdPid;
     const std::string etcdAddr = "127.0.0.1:2366,127.0.0.1:2368";
 };
@@ -55,6 +58,7 @@ TEST_F(EtcdClientTest, GetEtcdClusterStatus) {
     ASSERT_EQ(0, client.GetEtcdClusterStatus(&leaderAddr, &onlineState));
     std::map<std::string, bool> expected = {{"127.0.0.1:2366", true},
                                             {"127.0.0.1:2368", false}};
+    ASSERT_EQ(expected, onlineState);
     ASSERT_EQ("127.0.0.1:2366", leaderAddr);
 
     // 宕掉节点
@@ -62,7 +66,9 @@ TEST_F(EtcdClientTest, GetEtcdClusterStatus) {
     std::this_thread::sleep_for(std::chrono::seconds(2));
     ASSERT_EQ(0, client.GetEtcdClusterStatus(&leaderAddr, &onlineState));
     expected = {{"127.0.0.1:2366", false}, {"127.0.0.1:2368", false}};
+    ASSERT_EQ(expected, onlineState);
     ASSERT_EQ("", leaderAddr);
+    running = false;
 
     // 空指针错误
     ASSERT_EQ(-1, client.GetEtcdClusterStatus(nullptr, &onlineState));
